Move `getPhysicalActions` from ASTUtils to GeneratorBase
