+++ org.lflang/src/org/lflang/ASTUtils.java
@@ -636,24 +636,6 @@ public class ASTUtils {
     public static List<Action> allActions(Reactor definition) {
         return ASTUtils.collectElements(definition, featurePackage.getReactor_Actions());
     }
-
-
-    /**
-     * Given a list of reactors, return a list of all physical actions.
-     * @param reactors Generator class definition.
-     */
-    public static List<Action> allPhysicalActions(Iterable<Reactor> reactors) {
-        ArrayList<Action> physicalActions = new ArrayList<Action>();
-        for (Resource resource : GeneratorUtils.getResources(reactors)) {
-            var actions = Iterables.filter(IteratorExtensions.toIterable(resource.getAllContents()), Action.class);
-            for (Action action : actions) {
-                if (Objects.equal(action.getOrigin(), ActionOrigin.PHYSICAL)) {
-                    physicalActions.add(action);
-                }
-            }
-        }
-        return physicalActions;
-    }
     
     /**
      * Given a reactor class, return a list of all its connections,
